Implement algorithm that find shortest distance between 2 users. Remove destructors in User and UserStore.
This algorith is based in Depth First Search. It pretend find the minimum distance (in the path) between two selected nodes in a unweighted and undirected graph. The root will be one of the two users selected. First we explore all friends of the root finding for a solution with cost 0. If we not find a solution we explore all solutions of cost 1, and the same idea unti explore the full "graph". The first soultion founded will be the best, because we explore first the better solutions. For develop the algorithm we use a queue for store the "nodes" in the current level and visit their later. We also use an set of visited nodes for avoid cycles. Finally the solution will be the depth of the "tree" that we are generating from the start until find the solution.
In the worst case this algorithm visit all "nodes"(users) one time, insert the "nodes" in the visited set one time, and visit all his "node" adjacents one time. In the worst case we explore all friends list with a cost of O(nodes+edges). Where nodes are users and edges are every friend. Then the time of execution will be determined for the time that we need for visit the "adjacents" or friends.

Remove the destructor for classes User and StoreUser because they are generating segmentation faults when the automatic objects invoque destructor after his life cycle.