Change the data structure that store User in class "StoreUsers"
Change the std::set by an hash table with the size of the number of user. Its implementation need more memory but the access to the table array is equal to O(1) more the cost of calculate the function hash. The cost of the function has is proportional to the string of size name. Every position of the array has one set of user but this set has now and smaller size. Now we are wasting a lot of memory unnecesary because, in the table initiation we are creating one std::set that will be probably empty.